refactor(feature/fundo-tf): Reorganizadas pastas e corrigido bug no stoploss
